pointer protos + restructure raycasts
