+++ src/dcl/mod.rs
@@ -37,7 +37,7 @@ pub enum RendererResponse {
 }
 
 #[allow(clippy::large_enum_variant)] // we don't care since the error case is very rare
-// data from scene to renderer
+                                     // data from scene to renderer
 pub enum SceneResponse {
     Error(SceneId, String),
     Ok(SceneId, SceneCensus, CrdtStore),
