+++ src/dcl_component/proto_components.rs
@@ -43,6 +43,8 @@ impl DclProtoComponent for sdk::components::PbMeshCollider {}
 impl DclProtoComponent for sdk::components::PbMaterial {}
 impl DclProtoComponent for sdk::components::PbGltfContainer {}
 impl DclProtoComponent for sdk::components::PbAnimator {}
+impl DclProtoComponent for sdk::components::PbPointerEvents {}
+impl DclProtoComponent for sdk::components::PbPointerEventsResult {}
 
 // VECTOR3 conversions
 impl Copy for common::Vector3 {}
